+++ components/clarity-repl/src/repl/datastore.rs
@@ -6,6 +6,7 @@ use clarity::types::chainstate::ConsensusHash;
 use clarity::types::chainstate::SortitionId;
 use clarity::types::chainstate::StacksAddress;
 use clarity::types::chainstate::StacksBlockId;
+use clarity::types::chainstate::TrieHash;
 use clarity::types::chainstate::VRFSeed;
 use clarity::types::StacksEpochId;
 use clarity::util::hash::Sha512Trunc256Sum;
@@ -36,14 +37,18 @@ fn epoch_to_peer_version(epoch: StacksEpochId) -> u8 {
         StacksEpochId::Epoch24 => PEER_VERSION_EPOCH_2_4,
         StacksEpochId::Epoch25 => PEER_VERSION_EPOCH_2_5,
         StacksEpochId::Epoch30 => PEER_VERSION_EPOCH_3_0,
+        StacksEpochId::Epoch31 => PEER_VERSION_EPOCH_3_1,
     }
 }
+
+#[derive(Clone, Debug)]
+struct StoreEntry(StacksBlockId, String);
 
 #[derive(Clone, Debug)]
 pub struct ClarityDatastore {
     open_chain_tip: StacksBlockId,
     current_chain_tip: StacksBlockId,
-    store: HashMap<StacksBlockId, HashMap<String, String>>,
+    store: HashMap<String, Vec<StoreEntry>>,
     metadata: HashMap<(String, String), String>,
     block_id_lookup: HashMap<StacksBlockId, StacksBlockId>,
     height_at_chain_tip: HashMap<StacksBlockId, u32>,
@@ -117,7 +122,7 @@ impl ClarityDatastore {
         Self {
             open_chain_tip: id,
             current_chain_tip: id,
-            store: HashMap::from([(id, HashMap::new())]),
+            store: HashMap::new(),
             metadata: HashMap::new(),
             block_id_lookup: HashMap::from([(id, id)]),
             height_at_chain_tip: HashMap::from([(id, 0)]),
@@ -140,7 +145,6 @@ impl ClarityDatastore {
     ///     blockhash would already have committed and no longer exist in the save point stack.
     /// this is a "lower-level" rollback than the roll backs performed in
     ///   ClarityDatabase or AnalysisDatabase -- this is done at the backing store level.
-
     pub fn begin(&mut self, _current: &StacksBlockId, _next: &StacksBlockId) {
         // self.marf.begin(current, next)
         //     .expect(&format!("ERROR: Failed to begin new MARF block {} - {})", current, next));
@@ -176,31 +180,26 @@ impl ClarityDatastore {
         //     .expect("ERROR: Failed to commit MARF block");
     }
 
-    pub fn put(&mut self, key: &str, value: &str) {
-        let lookup_id = self
-            .block_id_lookup
-            .get(&self.open_chain_tip)
-            .expect("Could not find current chain tip in block_id_lookup map");
-
-        // if there isn't a store for the open chain_tip, make one and update the
-        // entry for the block id in the lookup table
-        if *lookup_id != self.open_chain_tip {
-            self.store.insert(
-                self.open_chain_tip,
-                self.store
-                    .get(lookup_id)
-                    .unwrap_or_else(|| panic!("Block with ID {:?} does not exist", lookup_id))
-                    .clone(),
-            );
-
-            self.block_id_lookup
-                .insert(self.open_chain_tip, self.current_chain_tip);
-        }
-
-        if let Some(map) = self.store.get_mut(&self.open_chain_tip) {
-            map.insert(key.to_string(), value.to_string());
+    fn put(&mut self, key: &str, value: &str) {
+        if let Some(entries) = self.store.get_mut(key) {
+            entries.push(StoreEntry(self.open_chain_tip, value.to_string()));
         } else {
-            panic!("Block does not exist for current chain tip");
+            self.store.insert(
+                key.to_string(),
+                vec![StoreEntry(self.open_chain_tip, value.to_string())],
+            );
+        }
+    }
+
+    fn get_latest_data(&self, data: &[StoreEntry]) -> Option<String> {
+        let StoreEntry(tip, value) = data.last()?;
+
+        if self.height_at_chain_tip.get(tip)?
+            <= self.height_at_chain_tip.get(&self.current_chain_tip)?
+        {
+            Some(value.clone())
+        } else {
+            self.get_latest_data(&data[..data.len() - 1])
         }
     }
 
@@ -219,16 +218,25 @@ impl ClarityBackingStore for ClarityDatastore {
 
     /// fetch K-V out of the committed datastore
     fn get_data(&mut self, key: &str) -> Result<Option<String>> {
-        let lookup_id = self
-            .block_id_lookup
-            .get(&self.current_chain_tip)
-            .expect("Could not find current chain tip in block_id_lookup map");
-
-        if let Some(map) = self.store.get(lookup_id) {
-            Ok(map.get(key).cloned())
-        } else {
-            panic!("Block does not exist for current chain tip");
+        match self.store.get(key) {
+            Some(data) => Ok(self.get_latest_data(data)),
+            None => Ok(None),
         }
+    }
+
+    fn get_data_from_path(&mut self, _hash: &TrieHash) -> Result<Option<String>> {
+        unreachable!()
+    }
+
+    fn get_data_with_proof(&mut self, _key: &str) -> Result<Option<(String, Vec<u8>)>> {
+        Ok(None)
+    }
+
+    fn get_data_with_proof_from_path(
+        &mut self,
+        _hash: &TrieHash,
+    ) -> Result<Option<(String, Vec<u8>)>> {
+        unreachable!()
     }
 
     fn has_entry(&mut self, key: &str) -> Result<bool> {
@@ -297,10 +305,6 @@ impl ClarityBackingStore for ClarityDatastore {
             Some(result) => Ok(Some(result.to_string())),
             None => Ok(None),
         }
-    }
-
-    fn get_data_with_proof(&mut self, _key: &str) -> Result<Option<(String, Vec<u8>)>> {
-        Ok(None)
     }
 
     fn get_contract_hash(