+++ CHANGELOG.md
@@ -1,17 +1,36 @@
-# [2.11.2](https://github.com/hirosystems/clarinet/compare/v2.11.2...v2.11.2) (2024-11-30)
+# [2.12.0](https://github.com/hirosystems/clarinet/compare/v2.11.2...v2.12.0) (2024-12-18)
+
+##### New Features
+
+*  Add costs report in sdk tx response (#1620) (6b68e00f)
+*  Epoch 3.1 support (#1630) (682daf54)
+
+##### Bug Fixes
+
+*  Typo for initEmptySession function (#1619) (1e731094)
+
+##### Performance Improvements
+
+*  Improve simnet datastore data struct (#1617) (ad66d1bd)
+
+##### Tests
+
+*  Setup basic wasm-bindgen tests for the sdk (#1614) (3f32f19c)
+
+# [2.11.2](https://github.com/hirosystems/clarinet/compare/v2.11.1...v2.11.2) (2024-10-30)
 
 ##### Bug Fixes
 
 *  Default devnet stacker (#1602) (af78d2d6)
 
-# [2.11.1](https://github.com/hirosystems/clarinet/compare/v2.11.0...v2.11.1) (2024-11-29)
+# [2.11.1](https://github.com/hirosystems/clarinet/compare/v2.11.0...v2.11.1) (2024-10-29)
 
 ##### Refactors
 
 *  Set epoch 3.0 block height (#1599) (98ef414f)
 *  Have a stacking order by default in devnet config (#1598) (2c24ac00)
 
-# [2.11.0](https://github.com/hirosystems/clarinet/compare/v2.10.0...v2.11.0) (2024-11-28)
+# [2.11.0](https://github.com/hirosystems/clarinet/compare/v2.10.0...v2.11.0) (2024-10-28)
 
 ##### New Features
 
@@ -22,7 +41,7 @@
 
 *  Update SignerMessage codec (#1591) (9e090c58)
 
-# [2.10.0](https://github.com/hirosystems/clarinet/compare/v2.9.0...v2.10.0) (2024-11-17)
+# [2.10.0](https://github.com/hirosystems/clarinet/compare/v2.9.0...v2.10.0) (2024-10-17)
 
 ##### New Features